Added trigger for fire
guys now gain back warmth while near fire

+++ strandtype/Assets/Player1/P1HealthBar.cs
@@ -8,13 +8,13 @@ public class P1HealthBar : MonoBehaviour
 
     public Slider slider;
 
-    public void P1MaxWarmth(int p1warmth)
+    public void P1MaxWarmth(float p1warmth)
     {
         slider.maxValue = p1warmth;
         slider.minValue = p1warmth;
     }
 
-    public void P1SetWarmth(int p1warmth)
+    public void P1SetWarmth(float p1warmth)
     {
         slider.value = p1warmth;
     }
