+++ strandtype/Assets/Player2/P2HealthBar.cs
@@ -8,13 +8,13 @@ public class P2HealthBar : MonoBehaviour
 
     public Slider slider;
 
-    public void P2MaxWarmth(int p2warmth)
+    public void P2MaxWarmth(float p2warmth)
     {
         slider.maxValue = p2warmth;
         slider.minValue = p2warmth;
     }
 
-    public void P2SetWarmth(int p2warmth)
+    public void P2SetWarmth(float p2warmth)
     {
         slider.value = p2warmth;
     }
